requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
gql==3.3.0
graphql-core==3.2.1
idna==3.3
multidict==6.0.2
python-dotenv==0.20.0
requests==2.28.0
requests-toolbelt==0.9.1
urllib3==1.26.9
yarl==1.7.2
